+++ NovaWing/NovaWing/Scene/ClearScene.cpp
@@ -555,8 +555,8 @@ void ClearScene::Draw()
 	//二つの画像を描画
 	//Aボタン画像
 	DrawRotaGraph(
-		wsize.width * a_button_pos.x,
-		wsize.height * a_button_pos.y,
+		static_cast<int>(wsize.width * a_button_pos.x),
+		static_cast<int>(wsize.height * a_button_pos.y),
 		a_button_scale, 0.0, aButtonHandle, true
 	);
 
@@ -565,8 +565,8 @@ void ClearScene::Draw()
 	{
 		//決定のテキスト画像
 		DrawRotaGraph(
-			wsize.width * decide_graph_pos.x,
-			wsize.height * decide_graph_pos.y,
+			static_cast<int>(wsize.width * decide_graph_pos.x),
+			static_cast<int>(wsize.height * decide_graph_pos.y),
 			decide_graph_scale, 0.0, decideHandle, true
 		);
 	}
@@ -574,8 +574,8 @@ void ClearScene::Draw()
 	{
 		//次へ のテキスト画像
 		DrawRotaGraph(
-			wsize.width * next_graph_pos.x,
-			wsize.height *next_graph_pos.y,
+			static_cast<int>(wsize.width * next_graph_pos.x),
+			static_cast<int>(wsize.height * next_graph_pos.y),
 			next_graph_scale, 0.0, nextHandle, true
 		);
 	}
@@ -592,35 +592,35 @@ void ClearScene::DrawResultText(int fontHandle)
 
 	//敵を倒した数
 	std::wstring killCountWString = std::to_wstring(static_cast<int>(m_currentKillCount));
-	int killCountWidth = GetDrawStringWidthToHandle(killCountWString.c_str(), killCountWString.size(), fontHandle);
+	int killCountWidth = GetDrawStringWidthToHandle(killCountWString.c_str(), static_cast<int>(killCountWString.size()), fontHandle);
 	DrawStringToHandle(
-		wsize.width * kill_count_pos.x - killCountWidth,
-		wsize.height * kill_count_pos.y,
+		static_cast<int>(wsize.width * kill_count_pos.x) - killCountWidth,
+		static_cast<int>(wsize.height * kill_count_pos.y),
 		killCountWString.c_str(), score_color, fontHandle
 	);
 	//クリアタイム
 	std::wstring clearTimeWString = std::to_wstring(static_cast<int>(m_currentClearTime));
-	int clearTimeWidth = GetDrawStringWidthToHandle(clearTimeWString.c_str(), clearTimeWString.size(), fontHandle);
+	int clearTimeWidth = GetDrawStringWidthToHandle(clearTimeWString.c_str(), static_cast<int>(clearTimeWString.size()), fontHandle);
 	DrawStringToHandle(
-		wsize.width * clear_time_pos.x - clearTimeWidth,
-		wsize.height * clear_time_pos.y,
+		static_cast<int>(wsize.width * clear_time_pos.x) - clearTimeWidth,
+		static_cast<int>(wsize.height * clear_time_pos.y),
 		clearTimeWString.c_str(), score_color, fontHandle
 	);
 	//被弾回数
 	std::wstring hitCountWString = std::to_wstring(static_cast<int>(m_currentHitCount));
-	int hitCountWidth = GetDrawStringWidthToHandle(hitCountWString.c_str(), hitCountWString.size(), fontHandle);
+	int hitCountWidth = GetDrawStringWidthToHandle(hitCountWString.c_str(), static_cast<int>(hitCountWString.size()), fontHandle);
 	DrawStringToHandle(
-		wsize.width * hit_count_pos.x - hitCountWidth,
-		wsize.height * hit_count_pos.y,
+		static_cast<int>(wsize.width * hit_count_pos.x) - hitCountWidth,
+		static_cast<int>(wsize.height * hit_count_pos.y),
 		hitCountWString.c_str(), hit_count_color, fontHandle
 	);
 
 	//スコア
 	std::wstring scoreWString = std::to_wstring(static_cast<int>(m_currentScore));
-	int scoreWidth = GetDrawStringWidthToHandle(scoreWString.c_str(), scoreWString.size(), fontHandle);
+	int scoreWidth = GetDrawStringWidthToHandle(scoreWString.c_str(), static_cast<int>(scoreWString.size()), fontHandle);
 	DrawStringToHandle(
-		wsize.width * score_pos.x - scoreWidth,
-		wsize.height * score_pos.y,
+		static_cast<int>(wsize.width * score_pos.x) - scoreWidth,
+		static_cast<int>(wsize.height * score_pos.y),
 		scoreWString.c_str(), score_color, fontHandle
 	);
 
@@ -632,26 +632,26 @@ void ClearScene::DrawResultText(int fontHandle)
 	ClearDrawScreen();
 	//敵を倒した数
 	DrawStringToHandle(
-		wsize.width * kill_count_pos.x - killCountWidth,
-		wsize.height * kill_count_pos.y,
+		static_cast<int>(wsize.width * kill_count_pos.x) - killCountWidth,
+		static_cast<int>(wsize.height * kill_count_pos.y),
 		killCountWString.c_str(), score_color, fontHandle
 	);
 	//クリアタイム
 	DrawStringToHandle(
-		wsize.width * clear_time_pos.x - clearTimeWidth,
-		wsize.height * clear_time_pos.y,
+		static_cast<int>(wsize.width * clear_time_pos.x) - clearTimeWidth,
+		static_cast<int>(wsize.height * clear_time_pos.y),
 		clearTimeWString.c_str(), score_color, fontHandle
 	);
 	//被弾回数
 	DrawStringToHandle(
-		wsize.width * hit_count_pos.x - hitCountWidth,
-		wsize.height * hit_count_pos.y,
+		static_cast<int>(wsize.width * hit_count_pos.x) - hitCountWidth,
+		static_cast<int>(wsize.height * hit_count_pos.y),
 		hitCountWString.c_str(), hit_count_color, fontHandle
 	);
 	//スコア
 	DrawStringToHandle(
-		wsize.width * score_pos.x - scoreWidth,
-		wsize.height * score_pos.y,
+		static_cast<int>(wsize.width * score_pos.x) - scoreWidth,
+		static_cast<int>(wsize.height * score_pos.y),
 		scoreWString.c_str(), score_color, fontHandle
 	);
